requirements.txt
python-dotenv
mysqlclient
django
kafka-python
pyyaml
sqlparse
python-magic
channels
asyncio
aiokafka
aiohttp
tushare
lxml
pandas
simplejson
requests
bs4
attrs

terminado
tornado
tornado-xstatic
XStatic
XStatic-term.js
